requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
paho-mqtt==1.5.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
